backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.3
openai==1.55.3
faiss-cpu==1.7.4
python-dotenv==1.0.0
pydantic==2.6.0
python-chess==1.999
aiohttp==3.9.1
langchain-community
langchain-openai 